+++ client/packages/grandpa-ranger/config/local.ts
@@ -2,11 +2,11 @@ export default {
 	circuit: {
 		rpc1: {
 			ws: "ws://localhost:9944",
-			http: "http://localhost:9933"
+			http: "http://localhost:9944"
 		},
 		rpc2: {
 			ws: "ws://localhost:9944",
-			http: "http://localhost:9933"
+			http: "http://localhost:9944"
 		},
 	},
 	target: { // we dont need to specify the http endpoint for the target
@@ -19,5 +19,4 @@ export default {
 	},
 	rangeInterval: 10, // time between range submissions in seconds
 	targetGatewayId: "roco",
-	circuitSigner: "0x0177d124e501887c2470e260c8f0da60db9ed3dba808a682f09afb39eff0c561"
 }